Move Pricing Provider
